Miller Rebuilt Reborn Reloaded
After tons of garbage, I got rid of a lot of bad stuff and recreated a
newer fresher repository.
This repo on this commit works on tomcat servers and looks nice. Throw
in your hibernate.cfg.xml file to build/classes and then you should be
good 2 go. I will leave the old build on github for reference but
further work will be done on MillerRebuilt (2016).