CloseoutEnhanced
CloseoutEnhanced is now a thing. A terrible terribly big thing. It’s a
huge wall of text, never look in there.
BUT
when I tested out the system I was able to get copSubmitted(Date,Notes,
and Status) into the database. I have not tried it with the rest but I
will tomorrow.
If I’m able to get everything into the database then I’ll finally be
able to go back tomorrow and get inspections and whatever else is new
onto closeoutenhanced. I should be able to do that and then it’s just
cleaning what I can (which won’t be too much because most of it is just
fields… over and over again…)