CloseoutEnhanced Completed
At least in the add projects and edit projects. But it’s done!
Unfortunately I kind of worked around the current system so next week
I’ll probably have to merge my work into the previous closeout system.
Which will be a lot of typing again… but it’ll make people happier
because with the system because that data won’t be essentially lost.
It’ll just take a little bit of time to put my content into the
existing content.