Goodbye CloseoutEnhanced
Moved everything to CloseoutDetails so that my additions are weaved
into the current system instead of being sloppily stapled on.
Everything works so far, I’m just trying to figure out how to get the
report to display all of the fields correctly. It’ll take some time to
enter all of the data and some time to actually figure out the best way
to format that data.