Closeout Reports
Reports have closeouts. It took a day but it happened. In order to get
specific closeout reports you have to use the parameters to get
specifically “inspections” or “liens” or “warranties” but all of the
basic data is included on the general report (along with general
closeout documents). Next up is to delete a third of all of the new
fields. Delete all of the comments and then add one for each category
of closeout (closeout docs, inspect, warrant, liens). It should take a
little bit only because it’s a lot of deleting. Also to delete the
entries for the SQL table! A clean table is a happy table;