Finished up closeoutData and started permitData
Permit data is weird. The closeout_id table entry automatically
incremented and created null values but permits and inspections didn’t
so only recent project get those id associated with them. This makes
working with old projects really strange and I still have to figure out
what to do. But it might be possible just to ignore it for the old
stuff and look to the future though that way is not entirely preferred…