Permit and Inspection Data Completed
Everything is starting to become separated with permits and inspections
joining the way of the closeout data (going to its own page). I cleaned
up a bunch of ProjectService as well because it probably is some of the
nastiest looking code I’ve ever seen. Knocked it down from 2 thousand
lines of code to only 1 thousand lines and by the time I’m finished
with everything it’ll probably be down to ~500 lines. - All the code is
being abstracted away to service helpers which do the line heavy tasks
that you simply don’t need to see in ProjectService.

Also I added SalvageValue to closeout data because we use that.