Basic Info Page
This one is kind of strange because the info is not encapsulated by
another object so I kept on getting exceptions. Nothing that can’t be
fixed so no worries at least.

Also continued work on minimizing the lines in ProjectService. To get
down to that ~500 line goal all that equipment stuff is going to have
to be dealt with.