Finally Change Orders
I got a lot of the groundwork in the system for the change orders and
now alls that’s left is to do (which won’t be necessarily easy) is to
get everything onto the system. It would be great to be able to just
use the preexisting ChangeOrder object to hold my information but I’m
not exactly sure what it’s connected to and don’t want to break
anything. I might just make a new object to hold an ID and an array of
change orders. We will have to see.