Printouts for ChangeOrders
Since we had no way to get the information out of the system I figured
that it would probably be beneficial for there to be a way… to do that…
So now we do! I haven’t tried to print it yet but I would assume that
it will print well enough that people can use it to print out documents
for projects they want to see.