Permit Time
It’s permit time now. Other old permit system is almost completely
being abandoned for this new form. I’m still creating it so that the
old data is integrated by just carrying the dates over to the new form
in the most appropriate field. But honestly, I’m pretty sure that we’ve
never even actually ever had data in the permits!