Users can find their projects
In order to facilitate new users coming in, each person is probably
going to be more interested in their projects than other people’s so I
provided a link that will automatically bring the person to the find
project page filtering for their active projects. Further parameters
can be specified as needed.