Most of Equipment
Most of the equipment stuff is in. I just need to send in the data with
an ajax call and then to create a addEquipment/editEquipment function.
Not much really but needs to be tested a bunch before actual
deployment.

I also got rid of a lot of unused project objects. We got
CloseoutDetails that has everything so *Closeout is pretty useless.

I didn’t even realize defects.html was still in the system. It’s gone.
Forever.

I’m hoping to at least get the base code done for equipment by
tomorrow. Testing will probably have to wait for later.