Trigger Bare Skeleton
Bare bones to a lot of the trigger system. The browser (which mimics
the find project page), trigger form (which will work a lot like the
reports generation page), and the backend code for it.

Currently we can trigger on attributes immediately within the Project
object. The only challenge will be finding out exactly how to create
criteria for Project’s foreign keys. But that shouldn’t be awful
whatsoever.