A lot more Alert/Trigger work
New trigger form almost complete. Just need to automate a few more
things in the individual param-field window and then need to fix the
buttons to add new parameters, remove parameters, and to finally create
the trigger. The hardest of those will be to create the trigger. The
most tedious though will definitely be adding a new parameter.

Then once that is complete, backend work can be finalized to create
triggers for the database and to then generate alerts based off of
those triggers (instead of just printing the amount of projects queried
by a trigger).