Trigger Retrievin
The trigger form is nearly fully functional now. The only thing that
needs to be changed is for the Add Parameter and Remove This Parameter
button to work.

Everything else is basically abstracted away in something.each(…)
blocks which should handle any new parameters being added. Though
currently instead of creating a list of parameters we just assign the
parameter list to the most recently read parameter item. This is a
simple change which I have inline the comments but is there so that the
server would work.

The server needs a bit of work to parse the request appropriately and
separate JSON parameter objects instead of just reading the one. Once
that is done, we will just have to iterate over each parameter to
create a param string for the Trigger object.

A perhaps solvable but possibly not worth it issue to solve is that a
trigger can only check one object’s fields. Meaning that if I want to
check the project.actualTurnover field, I cannot also check the
project.closeoutDetails.buildingFinalPermitDate field as it belongs to
another object. We could simply couple fields with objects but it might
be more than that and at that point it might not be worth it. Users
might not need to compare permit fields against closeout fields
(basically boiling down to subset data vs subset data which shouldn’t
be common and therefore passable).

Mostly server stuff to do from here.