autofill closeout
If class = "AIA", then autofills punch list, as built drawings, and closeout photos with "Required". Otherwise autofills them with "N/A".  All warranty letter, lien, and AIA/MG2 fields are autofilled to "TBD". If permits = "yes" or "TBD", the final inspection fields are autofilled to "TBD". Otherwise they are autofilled to "N/A"