Scorecard changes and CSS changes
Added the Back to Failed Rules functionality to the scorecard. Made some changes to CSS where it was not properly adjusted within the div.  Changed the Save Changes button to green and aligned them right as per Bua's request. Changed the sort functionality for MCS number to descending from ascending.  Added new rules to Permits & Inspections and it looks like Permits & Inspections rules section is done. Also, in the remote desktop, added a new task in the task scheduler to the already existing tasks so that server restart is done automatically every moring at 5:00 AM