Scorecard 1.0. Updated database with PermitNotes 1000
Looks like Score Card version 1 is complete
Changed the rules in close out by appending the category to the end of the failed rule so that it can be used to go to that section while editing the score card rules.
Also updated the back to failed rules button such that making it a single button in each section, instead of a seperate button for example, say a seperate button in general info, scheduling and financial.
Modified the send Task Alerts via email functionality. The previous version has not been working before I joined for some reason, mostly because the host was specified wrong after that update by Justin. Now it works.
Also, the server was crashing someties when permits is saved because the character count or length of the permitNotes column in the databse is 255. Updated that too 1000 characters in the database.

Also, I have made some minor css changes