Changes in the description
Updated the code such that the white spaces are removed while updating or adding a task only if they are leading or trailing white spaces. This was the reason task report etc. was crashing because say for the name Alex Wagner, previously it was converting it to AlexWagner and this name is not present in the database and hence it's crashing.

Added notes and assignee name in the task alert email and reordered it according to Bua.

Also, made some minor CSS changes like making width 100% and adjusting the save changes buttons.