Task Email Alert when MCS person changed added
Added countChangeOrders method to count the number of MCS change orders and number of MCS change orders completed. Using the values from this function, the respective values are being updated in the table closeoutdetails. The reason for doing this is before, in closeout, javascript was calculationg the tasks and completed tasks and updating the number in the database only when save changes is clicked within closeout. This is sometinh wrong. What if the user changes the chage orders and it does not reflect in the closeoutdetails table in the database. This was affecting the scorecard directly, hence improved.

Updated the css for tasks by placing the buttons in the top right corner.

Edit task,  Edit Change Orders, Edit Equipment all these buttons are disabled in the beginning but once something in the table is selected and you return to find another projects, the edit task etc. buttons are now not disabled for this new project as it is getting the state from the previous project. This shouldn't be the case. Hence, I have disabled these buttons when you return to project manager or find another project.

Task alert is working good when a new task is added. However, Andy also wanted an alert if the MCS person is changed. To do this, I had to get the task at hand so I have created an action getSpecifictask. An ajax call at the beginning of submitTask() in taskform.html solved this.