Some Minor Updates. Commenting AJAX countChangeOrders call
Changed the sequence in Permits & Inspections as per Andy's Instructions. Also, updated the autofill for Permits & Inspections such that Yes-> Preparing and No -> N/A.
Added the chosen-sprite.png image to the themes in CSS.
Commenting the Count Change Orders AJAX call now because it was showing up a server error for Marvin. I'll keep it that way until I figure out why it is happening.
Added window.print() in print change orders as per Bua's instruction.