Added Start Date Column to projects table. Read bottom description for previous commit changes
Dave wanted a start date sorting in the intial projects page. So I have added a start date column to that and then made the respective changes like how I have done for mediumScore for ScoreCard. Here I have Used scheduledStart Date ................................................................................... Also, In the previous commit, I forgot to mention that I have changed the permits/Inspections div to a new one in which the dropdowns can be accessed directly