Updating Task, Changeorder, closeout, scheduling and financial rules
Scheduling rule assigned on sept 17th resolved.  Financial discrepancies ShouldInvoice and ActualInvoice are available only when the project status is closeout.  Also updated the rule for change order such that the rules are evaluated only if the change order status is not preparing. If it is preparing, added a new rule saying "Need to complete change order proposal and submit". Earlier, if a task if closed, some rules would still be active for that task. Updated that such that task task rules will be evaluated only if the task is not closed.